Do not suppress exceptions during expression evaluation
In ExpressionInterpreter, when evaluating expressions,
some of the exceptions were caught and wrapped in
failure function instead of causing immediate failure.
This behavior was intended for expression optimization
so as to delay the failure in hope that  the erroneous
expression is eventually removed from the plan.

This was not consistent though. Exceptions were caught
for CASE, IF and COALESCE expressions, while a simple
arithmetic expression `0/0` caused immediate failure.

After this change, the behavior remains the same for
`ExpressionInterpreter.optimize()`, but exceptions
are no more suppressed for `ExpressionInterpreter.evaluate()`.
The `evaluate()` method can be considered the actual
execution of the expression, e.g. ValuesNode execution
is based on `evaluate()`. If the expression should fail,
it should be no more delayed.